Change Customer.phone to type long

+++ src/main/java/babank/Customer.java
@@ -10,12 +10,12 @@ public class Customer extends User implements Serializable {
 	 * 
 	 */
 	private static final long serialVersionUID = -3980819666127803352L;
-	int phone;
+	long phone;
 	HashSet<Integer> accounts; // This will contain a list of all the account IDs that the user owns
 	
 	
 	
-	public Customer(String userName, String password, String name, UserType type, int phone) {
+	public Customer(String userName, String password, String name, UserType type, long phone) {
 		super(userName, password, name, type);
 		this.phone = phone;
 		accounts = new HashSet<Integer>();
